Extend parser to automatically handle NEXUS
The toplevel parser is now `p_tree = p_nexus ^ p_newick`. That is, it
first tries to parse the tree as a nexus file, if that fails, it parses
it as a newick. There is little performance cost, since the p_nexus
parser will fail on the first character of a newick file where it
finds '(' not '#' (from '#NEXUS').

Adding NEXUS was pretty easy, since NEXUS stores trees in NEWICK format.
Actually loading all the metadata from the NEXUS file would be more
complicated. Also adding PhyloXML and other formats would require
considerable work.

The nexus parser keeps only the first observed tree in the nexus file. I
may add handling for multi-tree files in the future if that becomes
desireable.