Reparameterize --scale=r to `n^(1/r)` rather than `n^r`
Change the `--scale=r` parameter in `sample para` and `sample mono` to
represent `n^(1/r)` instead of `n^r`. With this parameterization, `r`
ranges from 1 to infinity and thus does not conflict with the `--p
roportion` term (which ranges from 0 to 1). They do overlap at
`--scale=1` and `--proportion=1`, but these are equivalent, both
sampling everything.